feat(app): swap on berachain using a call on union
